Factors InMemoryTimerInternals out of ReduceFnTester
